Addressed comments.
