+++ slsim/Sources/SupernovaeCatalog/supernovae_sample.py
@@ -7,40 +7,8 @@ import numpy as np
 from astropy import units
 from scipy import stats
 from slsim.Sources.galaxies import galaxy_projected_eccentricity
-from lenstronomy.Util.param_util import transform_e1e2_product_average
-from lenstronomy.Util.param_util import ellipticity2phi_q
 from slsim.Util.param_util import ellipticity_slsim_to_lenstronomy
-
-
-def elliptical_distortion_product_average(x, y, e1, e2, center_x, center_y):
-    """Maps the coordinates x, y with eccentricities e1, e2 into a new elliptical
-    coordinate system with same coordinate orientation.
-
-    # TODO: Remove function after it is included in Lenstronomy
-
-    :param x: x-coordinate
-    :param y: y-coordinate
-    :param e1: eccentricity
-    :param e2: eccentricity
-    :param center_x: center of distortion
-    :param center_y: center of distortion
-    :return: distorted coordinates x', y'
-    """
-    x_, y_ = transform_e1e2_product_average(x, y, e1, e2, center_x, center_y)
-
-    # Rotate back
-    phi_g, q = ellipticity2phi_q(e1, e2)
-    cos_phi = np.cos(-phi_g)
-    sin_phi = np.sin(-phi_g)
-
-    x__ = cos_phi * x_ + sin_phi * y_
-    y__ = -sin_phi * x_ + cos_phi * y_
-
-    # Shift
-    x___ = x__ + center_x
-    y___ = y__ + center_y
-
-    return x___, y___
+from slsim.Util.param_util import elliptical_distortion_product_average
 
 
 def supernovae_host_galaxy_offset(host_galaxy_catalog):
@@ -72,8 +40,6 @@ def supernovae_host_galaxy_offset(host_galaxy_catalog):
     e2 = []
     transformed_x_off = []
     transformed_y_off = []
-    lenstronomy_e1 = []
-    lenstronomy_e2 = []
 
     for i in range(len(host_galaxy_catalog)):
 
@@ -112,8 +78,6 @@ def supernovae_host_galaxy_offset(host_galaxy_catalog):
         # Transform the offset coordinates with eccentricities e1, e2 into elliptical coordinate
         # system
         lens_e1, lens_e2 = ellipticity_slsim_to_lenstronomy(slsim_e1, slsim_e2)
-        lenstronomy_e1.append(lens_e1)
-        lenstronomy_e2.append(lens_e2)
 
         transform_x_off, transform_y_off = elliptical_distortion_product_average(
             x_off.value, y_off.value, lens_e1, lens_e2, 0, 0
